refactor: implement all DOM generation by script
