+++ slider.html
@@ -7,21 +7,6 @@
     <link rel="stylesheet" href="css/slider.css" />
   </head>
   <body>
-    <h1>Slider</h1>
-    <!-- <div class="slider-wrapper">
-      <div class="slider">
-        <button class="button-slider prev">
-          <span class="button-content">&#60;</span>
-        </button>
-        <div class="slider-window">
-          <div class="slider-line" style="transform: translateX(0px);"></div>
-        </div>
-        <button class="button-slider next">
-          <span class="button-content">&#62;</span>
-        </button>
-      </div>
-      <div class="dot-wrapper"></div>
-    </div> -->
     <script src="js/slider.js" type="module"></script>
   </body>
 </html>
